+++ src/lambda/local/detectLocalLambdas.ts
@@ -17,7 +17,6 @@ export interface LocalLambda {
     resource: CloudFormation.Resource
     templatePath: string
     handler?: string
-    codeUri?: string
 }
 
 export async function detectLocalLambdas(
@@ -72,8 +71,7 @@ async function detectLambdasFromTemplate(
             templatePath,
             protocol: getDebugProtocol(resources[key]!),
             handler: getHandler(resources[key]!),
-            resource: resources[key]!,
-            codeUri: getCodeUri(resources[key]!)
+            resource: resources[key]!
         }))
 }
 
@@ -98,14 +96,6 @@ function getDebugProtocol(resource: CloudFormation.Resource): 'inspector' | 'leg
 function getHandler(resource: CloudFormation.Resource): string | undefined {
     if (resource.Properties && resource.Properties.Handler) {
         return resource.Properties.Handler
-    }
-
-    return undefined
-}
-
-function getCodeUri(resource: CloudFormation.Resource): string | undefined {
-    if (resource.Properties && resource.Properties.CodeUri) {
-        return resource.Properties.CodeUri
     }
 
     return undefined
